feat: refactor CameraConfig as module for better readability and maintenance
